[MODIFY] adder.py|library.py Adding file for the adding functions and file for the common functions
