+++ makefile_completer.py
@@ -2,14 +2,11 @@
 
 import time
 import os
-import re
 import scripts.find_makefile as finder
 import scripts.definers as define
+import scripts.adder as add
 from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
-
-
-
 
 class FileEvent(FileSystemEventHandler):
     def on_created(self, event):
@@ -18,7 +15,7 @@ class FileEvent(FileSystemEventHandler):
             print("++ Path where the Makefile is: {}".format(dir_path))
             print("++ The file is in {} directory".format(define.PATHS[identifier]))
             if "Makefile" in os.listdir(dir_path):
-                addFiletoMakefile(event.src_path, dir_path, identifier)
+                add.addFiletoMakefile(event.src_path, dir_path, identifier)
             else:
                 print("++ Makefile not found in %s" % dir_path)
             print("\n++++++++ THE FILE WAS ADDED ++++++++\n\n\n")
@@ -49,50 +46,10 @@ class FileEvent(FileSystemEventHandler):
             print("==Path where the Makefile is: {}".format(dir_path))
             print("==The file is in {} directory".format(define.PATHS[identifier]))
             if "Makefile" in os.listdir(dir_path):
-                addFiletoMakefile(event.dest_path, dir_path, identifier)
+                add.addFiletoMakefile(event.dest_path, dir_path, identifier)
             else:
                 print("== Makefile not found int %s" % dir_path)
             print("\n======== THE FILE WAS MOVED =========\n\n\n")
-
-
-def readWrite(path, writing=None):
-    if writing == None:
-        with open(path, 'r', encoding='utf-8') as Makefile:
-            makeContent = Makefile.read()
-        return makeContent
-    with open(path, 'w', encoding='utf-8') as Makefile:
-        Makefile.write(writing)
-
-
-def addFiletoMakefile(realPath, dir_path, identifier):
-    makeContent = readWrite(path=dir_path + "Makefile", writing=None)
-    line_list = makeContent.splitlines()
-
-    print("++++++ ADDING ++++++\n")
-    fileName = realPath.replace(dir_path + ((define.IDENTIFIERS[identifier] + '/') if identifier != 2 else ''), '')
-    print("++ The file has de name of: {}".format(fileName))
-    fileLineMakefile = define.PATHS[identifier] + fileName + "\t\\"
-    print("++ The expected line is: {}\n".format(fileLineMakefile))
-
-    for line in line_list:
-        print(
-            "++ FINDING Makefile Line: {}\t||\tto find -> {}".format(line, fileLineMakefile))
-        if re.search(fileLineMakefile + "\\", line):
-            return
-
-    for i, line in enumerate(line_list):
-        print("++ TO ADD Makefile line: {}\t||\tto find -> {}".format(line, fileLineMakefile))
-        if re.match(define.MAKEFILE_VAR[identifier] + define.PATHS[identifier], line) or re.match(define.MAKEFILE_VAR[identifier] + "\t", line):
-            print("++ The line starts here: {}".format(line))
-            while line_list[i] != '':
-                i += 1
-            line_list.insert(i, fileLineMakefile)
-            print("++ {} was added !\n\n".format(fileLineMakefile))
-            break
-
-    makeContent = '\n'.join(line_list)
-    readWrite(path=dir_path + "Makefile", writing=makeContent)
-    print("++ Line well added\n\n")
 
 
 def removeFilefromMakefile(realPath, dir_path, identifier):
